requirements.txt
Django==3.2.8
django-allauth~=0.42.0
django-filter==21.1
djangorestframework==3.12.4
dj-database-url==0.5.0
dj-rest-auth==1.1.0
drf-yasg==1.20.0
flake8===3.9.2
gunicorn==20.1.0
model-bakery==1.3.3
psycopg2-binary==2.9.1
pytest==6.2.4
pytest-django==4.4.0
pytest-cov==2.12.1
pytest-mock==3.7.0
requests==2.27.1
whitenoise==5.3.0
